+++ frontend/src/components/Blocks/Schedule/Body.jsx
@@ -7,9 +7,9 @@ import { injectIntl } from 'react-intl';
 import { Schedule } from '@package/components';
 
 const Body = (props) => {
-  const { data } = props;
+  const { data, isEditMode } = props;
 
-  return <Schedule miscEvent={data.miscEvent} />;
+  return <Schedule miscEvent={data.miscEvent} isEditMode={isEditMode} />;
 };
 
 export default injectIntl(Body);